FINALLY

+++ Homework_25.py
@@ -1,3 +1,6 @@
+from operator import add
+
+
 a = {1,1,1,1,1,3,2,4,5,6}
 b = {10,11,12,13}
 # a.discard(1)
@@ -20,10 +23,12 @@ b = {10,11,12,13}
 # in lists popitem DNE
 # in set popitem DNE
 #in dict popitem removes last element and doesnt take any args
-num_1 = input('\nEnter first number: ')
-num_2 = input('\nEnter second number: ')
-s = int(num_1) + int(num_2)
+num_1 = float(input('\nEnter first number: '))
+num_2 = float(input('\nEnter second number: '))
+s = (num_1) + (num_2)
 print(s)
+#                                                               OR
+# print(add(num_1,num_2))
 
 
 # if num_1.isdigit() == True and num_2.isdigit() == True:
@@ -33,7 +38,7 @@ print(s)
 #if we don't use int class then we can use this but without int we can't add these two inputs
 #x = input('\nEnter number: ')
 x = input('enter a number: ')
-rem = int(x) % 2
+rem = float(x) % 2
 if rem == 1:
     print('entered number is odd')
 elif rem == 0:
